embedding-service/requirements.txt
# PlaceMates Embedding Microservice — Dependencies
# Python 3.10+

# ── Web Framework ──────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# ── ML / Embeddings ────────────────────────────────────────
sentence-transformers==3.3.1
torch>=2.1.0
transformers>=4.36.0

# ── Vector Search ──────────────────────────────────────────
faiss-cpu==1.9.0.post1

# ── ChromaDB (resume corpus) ──────────────────────────────
chromadb==0.5.23

# ── Utilities ──────────────────────────────────────────────
numpy>=1.24.0
